+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## [4.0.0]
 
 - **Dropped support for Web**
+- Changed type of `accessToken` and `idToken` in `OidcToken` to String
 - Updated min Dart version to 3.6.0
 - Updated Flutter version to 3.27.1
 - Updated all dependencies to latest versions